Modify build-tool to Gradle and dependent libraries to latest version.

+++ src/main/java/org/springframework/social/salesforce/connect/SalesforceOAuth2Template.java
@@ -1,9 +1,9 @@
 package org.springframework.social.salesforce.connect;
 
+import java.util.Map;
+
 import org.springframework.social.oauth2.AccessGrant;
 import org.springframework.social.oauth2.OAuth2Template;
-
-import java.util.Map;
 
 /**
  * Salesforce OAuth2Template implementation.
@@ -27,7 +27,7 @@ public class SalesforceOAuth2Template extends OAuth2Template {
 
 
     @Override
-    protected AccessGrant createAccessGrant(String accessToken, String scope, String refreshToken, Integer expiresIn, Map<String, Object> response) {
+    protected AccessGrant createAccessGrant(String accessToken, String scope, String refreshToken, Long expiresIn, Map<String, Object> response) {
         this.instanceUrl = (String) response.get("instance_url");
 
         return super.createAccessGrant(accessToken, scope, refreshToken, expiresIn, response);
